move rpc related code
